+++ mne/channels/channels.py
@@ -12,11 +12,11 @@ import numpy as np
 from scipy.io import loadmat
 from scipy import sparse
 
-from .externals.six import string_types
+from ..externals.six import string_types
 
-from .utils import verbose, logger
-from .io.pick import channel_type, pick_info
-from .io.constants import FIFF
+from ..utils import verbose, logger
+from ..io.pick import channel_type, pick_info
+from ..io.constants import FIFF
 
 
 def _get_meg_system(info):
@@ -92,10 +92,10 @@ def equalize_channels(candidates, verbose=None):
 
     Note. This function operates inplace.
     """
-    from .io.base import _BaseRaw
-    from .epochs import Epochs
-    from .evoked import Evoked
-    from .time_frequency import AverageTFR
+    from ..io.base import _BaseRaw
+    from ..epochs import Epochs
+    from ..evoked import Evoked
+    from ..time_frequency import AverageTFR
 
     if not all([isinstance(c, (_BaseRaw, Epochs, Evoked, AverageTFR))
                 for c in candidates]):
@@ -176,10 +176,10 @@ class PickDropChannelsMixin(object):
 
     def _pick_drop_channels(self, idx):
         # avoid circular imports
-        from .io.base import _BaseRaw
-        from .epochs import Epochs
-        from .evoked import Evoked
-        from .time_frequency import AverageTFR
+        from ..io.base import _BaseRaw
+        from ..epochs import Epochs
+        from ..evoked import Evoked
+        from ..time_frequency import AverageTFR
 
         if isinstance(self, (_BaseRaw, Epochs)):
             if not self.preload:
@@ -283,6 +283,10 @@ def _recursive_flatten(cell, dtype):
 def read_ch_connectivity(fname, picks=None):
     """Parse FieldTrip neighbors .mat file
 
+    More information on these neighbor definitions can be found on the
+    related FieldTrip documentation pages:
+    http://fieldtrip.fcdonders.nl/template/neighbours
+
     Parameters
     ----------
     fname : str
@@ -298,9 +302,9 @@ def read_ch_connectivity(fname, picks=None):
     """
     if not op.isabs(fname):
         templates_dir = op.realpath(op.join(op.dirname(__file__),
-                                            'neighbors'))
+                                            'data', 'neighbors'))
         templates = os.listdir(templates_dir)
-        if not any(f not in templates for f in (fname, fname + '.mat')):
+        if not any(f in templates for f in (fname, fname + '.mat')):
             raise ValueError('I do not know about this neighbor '
                              'template: "{}"'.format(fname))
         else: