+++ mne/channels/data/neighbors/__init__.py
@@ -1,0 +1,6 @@
+# Neighbor definitions for clustering permutation analysis.
+# This is a selection of files from http://fieldtrip.fcdonders.nl/template
+# Additional definitions can be obtained through the FieldTrip software.
+# For additional information on how these definitions were computed, please
+# consider the related fieldtrip documentation:
+# http://fieldtrip.fcdonders.nl/template/neighbours.