+++ mne/neighbors/__init__.py
@@ -1,3 +1,0 @@
-# Neighbor definitions for clustering permutation analysis.
-# This is a selection of files from https://github.com/fieldtrip/fieldtrip
-# Additional definitions can be obtained through the FieldTrip software.